Estrutura de FASES + pipeline de shaders + IA que evolui + portao de validacao
ESTRUTURA (o jogo virou fases, nao mundo continuo)
- biomeAtWorld/getRegionAt devolvem a zona da FASE: o mundo inteiro tem um bioma
  so. Antes o layout 3x3 se repetia ao infinito e o tema trocava enquanto se
  andava - nao existia "passei de fase", so um tapete continuo.
- Portal de fase: HUD com abates ate o portal, portal 3D (aros + coluna de
  energia + luz), [E] avanca com tela de transicao e recompensa (cura, creditos,
  equipamento por tier). A cada 3 fases o portal so abre com o CHEFE morto.
- Barreira de energia fecha o mundo da fase (owPhaseRadius); cenario e chunks
  param nela. content/phases.txt define a campanha inteira SEM recompilar.

RENDER (nao havia nenhum shader antes)
- resources/shaders: bright-pass -> blur separavel 1/4 -> ACES + contraste;
  world.vs/fs = half-lambert direcional (sol gira com o ciclo do dia), ambiente
  por zona, rim light e nevoa ate a cor do horizonte, aplicado nos modelos.
- Ceu tingido pela zona (era quase preto: mundo parecia colado no vazio).
- Chao dessaturado + contorno grosso no ator (legibilidade por silhueta).
- Ruas como GEOMETRIA (asfalto, meio-fio, faixa apagada, buracos) em cidade.
- Arquitetura POR BIOMA: predio moderno, bunker, cripta, espira, monolito,
  cabana, torre - e nada de castelo medieval em rua com asfalto.
- Escala ancorada em MEDIDA (log VOXSIZE: heroi = 66u), nao em chute.
- Z-fighting: camadas do chao separadas por decimos de unidade (piscava tudo).

ANIMACAO
- 4 poses de caminhada por tipo, capturadas do render 2D com a fase forcada;
  drawVoxel escolhe o quadro. Antes o modelo era congelado: todo mundo deslizava.
- NPC ganhou animacao de pernas (nao tinha nem no 2D).
- Chave do voxel do player inclui o equipamento: trocar de item muda o 3D.

IA
- EnemyDirector: observa distancia, kite, ritmo de abate e dano sofrido; muda a
  COMPOSICAO do spawn por arquetipo e distribui papeis (frontal/flanco/cerco).
  adaptLevel sobe quando o jogador domina e desce quando ele apanha.
- Tabelas de spawn separadas por fase (fantasma, ciborgue, morto-vivo no lugar).

QA
- BotController::passed() + validate.sh: pass/fail com motivo, vira exit code.
  Pegou nesta sessao: 4 travamentos, combate zerado 2x e livelock do escape.
- --test-seconds=N (o relatorio nunca saia) e --seed=N (mundo reprodutivel).

AUDIO
- Cadeia nova: filtro 2 polos, reverb de Schroeder, unissono desafinado,
  sub-grave, compressor e saturacao. Perfil acustico por zona.

+++ src/BuildingSystem.cpp
@@ -703,7 +703,7 @@ void BuildingSystem::renderArkBuilding(const Building& b) const {
     if (!b.built) {
         // Under construction scaffold
         DrawRectangleLinesEx({b.position.x - 36, b.position.y - 36, 72, 72}, 2.f, GRAY);
-        float prog = (int)(b.buildTimer * 72.f);
+        float prog = b.buildTimer * 72.f;   // era (int) num float: truncava e gerava C4244
         DrawRectangle((int)(b.position.x - 36), (int)(b.position.y + 28), (int)prog, 8, YELLOW);
         DrawText("Construindo...", (int)(b.position.x - 36), (int)(b.position.y - 50), 12, YELLOW);
         return;
